api/requirements.txt
pandas>=1.5.3, <1.6.0
uvicorn>=0.21.0, <0.22.0
fastapi>=0.94.1, <0.95.0
pydantic>=1.10.6, <1.11.0
loguru>=0.6.0, <0.7.0
pyarrow>=11.0.0, <11.1.0
gcsfs==2023.3.0
python-dotenv==0.21.1

